Added the jest config option watchPathIgnorePatterns because it was causing the build to fail on linux for android. ## Reason for Pull Request
- [ ] Bug fix
- [x] Feature/Change request

## Link for jira ticket
[CM-2303](http://evernym.atlassian.net/browse/CM-2303)

## In case of Feature/Change request

### High level description of changes done

- Upgrading react native to 0.58.6

## Tests written for Bug/Feature/Refactoring

- `<Please write test file/test name here>`

## Checklist

- [ ] I have created new screen, and I have checked the header of new screen on Android. I have also set the status bar color in app.js
- [ ] I have checked the back button functionality required for the screens in Android. I have added the routeNames that are need to be handled for the back button in app.js
- [ ] I have checked the back behaviour by swiping left or swiping down on screens and they behave as expected. Back behaviour of these screens has been approved by product team

+++ app/store/user/user-store.js
@@ -1,7 +1,7 @@
 // @flow
 import { put, takeLatest, call, all, select, take } from 'redux-saga/effects'
 import type { Saga } from 'redux-saga'
-import RNFetchBlob from 'react-native-fetch-blob'
+import RNFetchBlob from 'rn-fetch-blob'
 import ImagePicker from 'react-native-image-crop-picker'
 
 import {
@@ -164,15 +164,15 @@ export function* saveUserSelectedAvatarSaga(
   const { name, extension } = getImageInfo(imagePath)
   try {
     const appDirectory = RNFetchBlob.fs.dirs.DocumentDir
-    const existingAvatarName = yield select(getUserAvatarName)
-    const userAvatarPath = `${appDirectory}/${existingAvatarName}`
+    const existingAvatarName: ?string = yield select(getUserAvatarName)
+    const userAvatarPath = `${appDirectory}/${existingAvatarName || ''}`
     const alreadyExist = yield call(RNFetchBlob.fs.exists, userAvatarPath)
 
     const imageId = uuid()
     const avatarName = `${imageId}.${extension}`
     const newImagePath = `${appDirectory}/${avatarName}`
 
-    if (alreadyExist) {
+    if (alreadyExist && existingAvatarName) {
       yield call(RNFetchBlob.fs.unlink, userAvatarPath)
     }
